fix: handle errors in getting pod exit code
Previously, if a pod is failed prior to the creation of the initContainer,
the pod actor would crash in getting the exit code while the state of the
container already transitioned to terminated. This caused the pod actor not
to properly clean up the container state inside the task handler and the
scheduler, which leads the task state to be stuck in Active.